[feature] create delivery note for rulfilled orders, [fixes][minor] update shopify settings
