+++ erpnext_shopify/utils.py
@@ -17,6 +17,9 @@ def get_country():
 def get_shopify_customers():
 	return get_request('/admin/customers.json')['customers']
 
+def get_address_type(i):
+	return ["Billing", "Shipping", "Office", "Personal", "Plant", "Postal", "Shop", "Subsidiary", "Warehouse", "Other"][i]
+
 def create_webhook(topic, address):
 	post_request('admin/webhooks.json', json.dumps({
 		"webhook": {
@@ -25,9 +28,6 @@ def create_webhook(topic, address):
 			"format": "json"
 		}
 	}))
-
-def get_address_type(i):
-	return ["Billing", "Shipping", "Office", "Personal", "Plant", "Postal", "Shop", "Subsidiary", "Warehouse", "Other"][i]
 
 def shopify_webhook(f):
 	"""
@@ -69,7 +69,6 @@ def webhook_handler():
 	from webhooks import handler_map
 	topic = frappe.local.request.webhook_topic
 	data = frappe.local.request.webhook_data
-	print topic
 	handler = handler_map.get(topic)
 	if handler:
 		handler(data)
@@ -82,7 +81,6 @@ def get_request(path):
 	settings = get_shopify_settings()
 	s = get_request_session()
 	url = 'https://{}:{}@{}/{}'.format(settings['api_key'], settings['password'], settings['shopify_url'], path)
-	print url
 	r = s.get(url)
 	r.raise_for_status()
 	return r.json()
@@ -91,7 +89,6 @@ def post_request(path, data):
 	settings = get_shopify_settings()
 	s = get_request_session()
 	url = 'https://{}:{}@{}/{}'.format(settings['api_key'], settings['password'], settings['shopify_url'], path)
-	print "data",json.dumps(data),url
 	r = s.post(url, data=json.dumps(data), headers={'Content-type': 'application/json'})
 	r.raise_for_status()
 	return r.json()
@@ -102,3 +99,27 @@ def delete_request(path):
 	url = 'https://{}:{}@{}/{}'.format(settings['api_key'], settings['password'], settings['shopify_url'], path)
 	r = s.delete(url)
 	r.raise_for_status()
+	
+def delete_webhooks():
+	webhooks = get_webhooks()
+	for webhook in webhooks:
+		delete_request("/admin/webhooks/{}.json".format(webhook['id']))
+
+def get_webhooks():
+	webhooks = get_request("/admin/webhooks.json")
+	return webhooks["webhooks"]
+	
+def create_webhooks():
+	settings = get_shopify_settings()
+	for event in ["orders/create", "orders/delete", "orders/updated", "orders/paid", "orders/cancelled", "orders/fulfilled", 
+					"orders/partially_fulfilled", "order_transactions/create", "carts/create", "carts/update", 
+					"checkouts/create", "checkouts/update", "checkouts/delete", "refunds/create", "products/create", 
+					"products/update", "products/delete", "collections/create", "collections/update", "collections/delete", 
+					"customer_groups/create", "customer_groups/update", "customer_groups/delete", "customers/create", 
+					"customers/enable", "customers/disable", "customers/update", "customers/delete", "fulfillments/create", 
+					"fulfillments/update", "shop/update", "disputes/create", "disputes/update", "app/uninstalled", 
+					"channels/delete", "product_publications/create", "product_publications/update", 
+					"product_publications/delete", "collection_publications/create", "collection_publications/update", 
+					"collection_publications/delete", "variants/in_stock", "variants/out_of_stock"]:
+					
+		create_webhook(event, settings.webhook_address)
